refactor: The `ShareListForm` component is introduced to handle the form submission and uses the `shareList` function from the Firebase API to add the existing user to the list. This feature enhances collaboration and sharing within the application.
